+++ packages/spacecat-shared-content-client/test/clients/content-client.test.js
@@ -486,8 +486,8 @@ describe('ContentClient', () => {
       await expect(client.updateRedirects([{ from: '/A', to: '/B' }, 'malformed-redirect'])).to.be.rejectedWith('Redirect must be an object');
       await expect(client.updateRedirects([{ from: '/A', to: '/B' }, { from: '', to: '/B' }])).to.be.rejectedWith('Redirect must have a valid from path');
       await expect(client.updateRedirects([{ from: '/A', to: '/B' }, { from: '/A', to: '' }])).to.be.rejectedWith('Redirect must have a valid to path');
-      await expect(client.updateRedirects([{ from: 'A', to: '/B' }, { from: '/A', to: '/C' }])).to.be.rejectedWith('Invalid redirect from path: A');
-      await expect(client.updateRedirects([{ from: '/A', to: 'B' }, { from: '/A', to: '/C' }])).to.be.rejectedWith('Invalid redirect to path: B');
+      await expect(client.updateRedirects([{ from: 'A', to: '/B' }, { from: '/A', to: '/C' }])).to.be.rejectedWith('Invalid Redirect from path: A');
+      await expect(client.updateRedirects([{ from: '/A', to: 'B' }, { from: '/A', to: '/C' }])).to.be.rejectedWith('Invalid Redirect to path: B');
       await expect(client.updateRedirects([{ from: '/A', to: '/B' }, { from: '/A', to: '/A' }])).to.be.rejectedWith('Redirect from and to paths must be different');
     });
     it('update success', async () => {
@@ -554,6 +554,69 @@ describe('ContentClient', () => {
       const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
       await client.updateRedirects(newRedirects);
       await expect(redirectsSdk.append).to.not.have.been.called;
+    });
+  });
+
+  describe('updateBrokenInternalLink', () => {
+    let client;
+
+    beforeEach(async () => {
+      client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await client.getPageMetadata('/test-path'); // This will initialize the rawClient
+    });
+
+    it('should throw an error if brokenLink is not an object', async () => {
+      await expect(client.updateBrokenInternalLink('/test-path', 'not-an-object')).to.be.rejectedWith('URL must be an object');
+    });
+
+    it('should throw an error if brokenLink is missing from path', async () => {
+      const brokenLink = { to: 'https://new-link' };
+      await expect(client.updateBrokenInternalLink('/test-path', brokenLink)).to.be.rejectedWith('URL must have a valid from path');
+    });
+
+    it('should throw an error if brokenLink is missing to path', async () => {
+      const brokenLink = { from: 'http://old-link' };
+      await expect(client.updateBrokenInternalLink('/test-path', brokenLink)).to.be.rejectedWith('URL must have a valid to path');
+    });
+
+    it('should throw an error if brokenLink from path is invalid', async () => {
+      const brokenLink = { from: 'invalid-url', to: 'https://new-link' };
+      await expect(client.updateBrokenInternalLink('/test-path', brokenLink)).to.be.rejectedWith('Invalid URL from path: invalid-url');
+    });
+
+    it('should throw an error if brokenLink to path is invalid', async () => {
+      const brokenLink = { from: 'http://old-link', to: 'invalid-url' };
+      await expect(client.updateBrokenInternalLink('/test-path', brokenLink)).to.be.rejectedWith('Invalid URL to path: invalid-url');
+    });
+
+    it('should update broken internal link for Google Drive', async () => {
+      const brokenLink = { from: 'http://old-link', to: 'https://new-link' };
+      client.rawClient.getDocument = sinon.stub().returns({
+        updateLink: sinon.stub().resolves({ status: 200 }),
+      });
+      await client.updateBrokenInternalLink('/test-path', brokenLink);
+      expect(client.rawClient.getDocument.calledOnceWith('/test-path')).to.be.true;
+      expect(client.rawClient.getDocument().updateLink.calledOnceWith('http://old-link', 'https://new-link')).to.be.true;
+    });
+
+    it('should update broken internal link for OneDrive', async () => {
+      const brokenLink = { from: 'http://old-link', to: 'https://new-link' };
+      client = ContentClient.createFrom(context, siteConfigOneDrive);
+      await client.getPageMetadata('/test-path'); // This will initialize the rawClient
+      client.rawClient.getDocument = sinon.stub().returns({
+        updateLink: sinon.stub().resolves({ status: 200 }),
+      });
+      await client.updateBrokenInternalLink('/test-path', brokenLink);
+      expect(client.rawClient.getDocument.calledOnceWith('/test-path.docx')).to.be.true;
+      expect(client.rawClient.getDocument().updateLink.calledOnceWith('http://old-link', 'https://new-link')).to.be.true;
+    });
+
+    it('should throw an error if updateLink fails', async () => {
+      const brokenLink = { from: 'http://old-link', to: 'https://new-link' };
+      client.rawClient.getDocument = sinon.stub().returns({
+        updateLink: sinon.stub().resolves({ status: 500 }),
+      });
+      await expect(client.updateBrokenInternalLink('/test-path', brokenLink)).to.be.rejectedWith('Failed to update link from http://old-link to https://new-link // [object Object]');
     });
   });
 });